interceptor: rename ocinterceptor to octrace. (#170)
